move library to stix20 directory
