+++ mitreattack/stix20/__init__.py
@@ -1,0 +1,2 @@
+from .MitreAttackData import MitreAttackData
+from .custom_attack_objects import StixObjectFactory, Matrix, Tactic, DataSource, DataComponent
