+++ mitreattack/stix20/custom_attack_objects.py
